Add user authentication, user model, and initial migrations
This commit adds automatic user authentication and two pieces of middleware for managing user sessions. It includes the creation of the initial migration files for Django, and introduces the MyUser model. This change also provides initial setup for the user login and registration functionalities such as corresponding HTML templates and Python scripts. The CSS stylings for the project are also included.